pyproject.toml
[project]
name = "notion-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "mcp-use>=1.3.11",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
